Fix ESC and Motor delta temps

+++ main.c
@@ -1472,9 +1472,10 @@ static void process_packet_vesc(unsigned char *data, unsigned int len) {
 				// Update delta message
 				log_message_esc_delta.dt = currentTime - log_message_esc.dt;
 				log_message_esc_delta.esc_id = esc_telemetry.vesc_id;
+				//TODO: understand why many of these casts are absolutely necessary
 				log_message_esc_delta.vin = (int)(esc_telemetry.v_in * 10) - (int)log_message_esc.vin;
-				log_message_esc_delta.motor_temp = (int)(esc_telemetry.temp_motor * 10) - log_message_esc.motor_temp;
-				log_message_esc_delta.mosfet_temp = (int)(esc_telemetry.temp_mos * 10) - log_message_esc.mosfet_temp;
+				log_message_esc_delta.motor_temp = (int)(esc_telemetry.temp_motor * 10) - (int)log_message_esc.motor_temp;
+				log_message_esc_delta.mosfet_temp = (int)(esc_telemetry.temp_mos * 10) - (int)log_message_esc.mosfet_temp;
 				log_message_esc_delta.duty_cycle = (int)(esc_telemetry.duty_now * 10) - log_message_esc.duty_cycle;
 				log_message_esc_delta.motor_current = (int)(esc_telemetry.current_motor * 10) - log_message_esc.motor_current;
 				log_message_esc_delta.battery_current = (int)(esc_telemetry.current_in * 10) - log_message_esc.battery_current;
